Pioneer engine update and change of year 2015-2016

+++ src/GeoPatchContext.h
@@ -1,4 +1,4 @@
-// Copyright © 2008-2015 Pioneer Developers. See AUTHORS.txt for details
+// Copyright © 2008-2016 Pioneer Developers. See AUTHORS.txt for details
 // Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
 #ifndef _GEOPATCHCONTEXT_H
@@ -35,8 +35,8 @@ private:
 	static inline int VBO_COUNT_MID_IDX() { return (4*3*(edgeLen-3)) + 2*(edgeLen-3)*(edgeLen-3)*3; }
 	//                                            ^^ serrated teeth bit  ^^^ square inner bit
 
-	static inline int IDX_VBO_LO_OFFSET(const int i) { return i*sizeof(unsigned short)*3*(edgeLen/2); }
-	static inline int IDX_VBO_HI_OFFSET(const int i) { return (i*sizeof(unsigned short)*VBO_COUNT_HI_EDGE())+IDX_VBO_LO_OFFSET(4); }
+	static inline int IDX_VBO_LO_OFFSET(const int i) { return i*sizeof(Uint32)*3*(edgeLen/2); }
+	static inline int IDX_VBO_HI_OFFSET(const int i) { return (i*sizeof(Uint32)*VBO_COUNT_HI_EDGE())+IDX_VBO_LO_OFFSET(4); }
 
 	static RefCountedPtr<Graphics::IndexBuffer> indices;
 	static int prevEdgeLen;
